インスタンス化のバグ修正
