+++ FunctionBinder/FunctionBinder/FunctionBinder/FunctionBinder.cpp
@@ -1,61 +1,27 @@
 ﻿#include <iostream>
-#include "Reader.h"
+#include "FunctionBinder.h"
 
 
-template<class T>
-struct list {
-	struct Node {
-		T     value;
-		Node* next;
-	};
-	Node* head = nullptr;
-	Node* tail = nullptr;
+struct Massege_t {
+	int data;
+};
 
-	~list() {
-		for (Node* it = head; it;) {
-			auto temp = it->next;
-			delete it;
-			it = temp;
-		}
-	}
-	void push_back(const T& r)
-	{
-		if (tail)
-			tail = tail->next = new Node{ r };
-		else  /// 要素無し
-			head = tail = new Node{ r };
-	}
+using FunctionBinder_t = FunctionBinder<void(Massege_t)>;
 
-	struct Iterator {
-		Node* p;
-		T& operator*() { return p->value; }
-		Iterator& operator++() { p = p->next; return *this; }
-		bool operator!=(const Iterator& r) { return p != r.p; }
-	};
-	Iterator begin() { return { head }; }
-	Iterator end()   {
-		if (tail)
-			return { tail->next };
-		else
-			return { head };
+struct Reader : private FunctionBinder_t {
+
+	void callback(Massege_t arg) override {
+		std::cout << arg.data << std::endl;
 	}
 };
 
 Reader reader;
+Reader reader1;
+Reader reader2;
 
 int main() {
-	_CrtSetDbgFlag(_CRTDBG_ALLOC_MEM_DF | _CRTDBG_LEAK_CHECK_DF);
-	list<int> l;
+	_CrtSetDbgFlag(_CRTDBG_ALLOC_MEM_DF | _CRTDBG_LEAK_CHECK_DF);  /// メモリリーク検出用
 
-	l.push_back(100);
-	l.push_back(101);
-	l.push_back(102);
-	l.push_back(103);
-	l.push_back(104);
-	for (auto& it : l)
-		std::cout << it << std::endl;
 
-	//std::cout << "call" << std::endl;
-	//FunctionBinder<void(Massege_t)>::bind({ 100 });
-	//std::cout << "end" << std::endl;
+	FunctionBinder_t::bind({ 100 });
 }
