+++ FunctionBinder/FunctionBinder/FunctionBinder/FunctionBinder.vcxproj
@@ -132,7 +132,6 @@
   <ItemGroup>
     <ClInclude Include="Base.h" />
     <ClInclude Include="FunctionBinder.h" />
-    <ClInclude Include="Reader.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
