+++ FunctionBinder/FunctionBinder/FunctionBinder/FunctionBinder.vcxproj.filters
@@ -26,8 +26,5 @@
     <ClInclude Include="Base.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="Reader.h">
-      <Filter>ヘッダー ファイル</Filter>
-    </ClInclude>
   </ItemGroup>
 </Project>
